requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
idna==3.3
multidict==6.0.2
mysql-connector-python==8.0.28
protobuf==3.19.4
psycopg2==2.9.3
typing_extensions==4.1.1
yarl==1.7.2